+++ ginstagram/admin.py
@@ -4,5 +4,5 @@ from django.contrib.auth.admin import UserAdmin
 # Register your models here.
 from .models import Users
 
-admin.site.register(Users)
+admin.site.register(Users, UserAdmin)
