+++ ginstagram/templates/ginstagram/profile.html
@@ -1,0 +1,1 @@
+{{userInfo}}
