+++ ginstagram/urls.py
@@ -3,5 +3,5 @@ from django.urls import include, path
 
 urlpatterns = [
     path('', views.main),
-    path('login', views.login),
+    path('<user_name>/', views.profile, name='profile'),
 ]
